requirements.txt
sounddevice
numpy
pyyaml